Add 'bdd' to write tests in business language that critique the product
This is a small gem; but the functionality it provides depends on its
integration with an another service: the Kubernetes API.

For this reason, the intent here is to write tests that:
* Cover also the integration with the Kubernetes API. If this
integration stops working, the tests must fail. The goal of these tests
is to "critique" the product.
* Speak "business language". IE: acceptance tests that talk about the
functionality provided to the end-user.
(Brian Marick's test quadrant)